VuplusKexec - check that Vu Multiboot files are installed before installing Multiboot - if not issue message

+++ lib/python/Screens/VuplusKexec.py
@@ -4,9 +4,10 @@ from Components.Console import Console
 from Components.Label import Label
 from Components.Sources.StaticText import StaticText
 from Components.SystemInfo import SystemInfo
+from Screens.MessageBox import MessageBox
 from Screens.Screen import Screen
 from Screens.Standby import QUIT_REBOOT, TryQuitMainloop
-
+from Tools.Directories import fileExists
 
 STARTUP = "kernel=/zImage root=/dev/%s rootsubdir=linuxrootfs0" % getMachineMtdRoot()					# /STARTUP
 STARTUP_RECOVERY = "kernel=/zImage root=/dev/%s rootsubdir=linuxrootfs0" % getMachineMtdRoot() 			# /STARTUP_RECOVERY
@@ -48,27 +49,31 @@ class VuplusKexec(Screen):
 	def RootInit(self):
 		self["actions"].setEnabled(False) # This function takes time so disable the ActionMap to avoid responding to multiple button presses
 		if SystemInfo["CanKexecVu"]:
-			self.title = _("Vu+ MultiBoot Initialisation - will reboot after 10 seconds.")
-			self["description"].text = _("Vu+ MultiBoot Initialisation - will reboot after 10 seconds.")
-			with open("/STARTUP", 'w') as f:
-				f.write(STARTUP)
-			with open("/STARTUP_RECOVERY", 'w') as f:
-				f.write(STARTUP_RECOVERY)
-			with open("/STARTUP_1", 'w') as f:
-				f.write(STARTUP_1)
-			with open("/STARTUP_2", 'w') as f:
-				f.write(STARTUP_2)
-			with open("/STARTUP_3", 'w') as f:
-				f.write(STARTUP_3)
-			with open("/etc/recovery", 'w') as f:
-				f.write(recovery)
-			print("[VuplusKexec][RootInit] Kernel Root", getMachineMtdKernel(), "   ", getMachineMtdRoot())
-			cmdlist = []
-			cmdlist.append("dd if=/dev/%s of=/zImage" % getMachineMtdKernel())						# backup old kernel
-			cmdlist.append("dd if=/usr/bin/kernel_auto.bin of=/dev/%s" % getMachineMtdKernel())	# create new kernel
-			cmdlist.append("mv /usr/bin/STARTUP.cpio.gz /STARTUP.cpio.gz")						# copy userroot routine
-			cmdlist.append("sync; sleep 1; sync; sleep 1; sync")
-			Console().eBatch(cmdlist, self.RootInitEnd, debug=True)
+			if fileExists("/usr/bin/kernel_auto.bin") and fileExists("/usr/bin/STARTUP.cpio.gz"):
+				self.title = _("Vu+ MultiBoot Inititialisation - will reboot after 10 seconds.")
+				self["description"].text = _("Vu+ MultiBoot Inititialisation - will reboot after 10 seconds.")
+				with open("/STARTUP", 'w') as f:
+					f.write(STARTUP)
+				with open("/STARTUP_RECOVERY", 'w') as f:
+					f.write(STARTUP_RECOVERY)
+				with open("/STARTUP_1", 'w') as f:
+					f.write(STARTUP_1)
+				with open("/STARTUP_2", 'w') as f:
+					f.write(STARTUP_2)
+				with open("/STARTUP_3", 'w') as f:
+					f.write(STARTUP_3)
+				with open("/etc/recovery", 'w') as f:
+					f.write(recovery)
+				print("[VuplusKexec][RootInit] Kernel Root", getMachineMtdKernel(), "   ", getMachineMtdRoot())
+				cmdlist = []
+				cmdlist.append("dd if=/dev/%s of=/zImage" % getMachineMtdKernel())						# backup old kernel
+				cmdlist.append("dd if=/usr/bin/kernel_auto.bin of=/dev/%s" % getMachineMtdKernel())	# create new kernel
+				cmdlist.append("mv /usr/bin/STARTUP.cpio.gz /STARTUP.cpio.gz")						# copy userroot routine
+				cmdlist.append("sync; sleep 1; sync; sleep 1; sync")
+				Console().eBatch(cmdlist, self.RootInitEnd, debug=True)
+			else:
+				self.session.open(MessageBox, _("[VuplusKexec][create Vu Multiboot environment] - Unable to complete, Vu+ Multiboot files missing"), MessageBox.TYPE_INFO, timeout=30)
+				self.close()
 		else:
 			self.close()
